refactor(#1431): split MCP tools_workflow + tools_inspection into sub-packages
Pure structural decomposition of the two largest non-protected MCP tool
files under umbrella #1427:

- src/scistudio/ai/agent/mcp/tools_workflow.py (884 LOC) -> tools_workflow/
  sub-package with 7 sub-modules; largest is read.py at 303 LOC.
- src/scistudio/ai/agent/mcp/tools_inspection.py (809 LOC) -> tools_inspection/
  sub-package with 6 sub-modules; largest is read.py at 398 LOC.

No behavior change. The legacy public import surface is preserved via
package __init__.py re-exports: every tool function, Pydantic envelope,
and internal helper that was reachable before the refactor is still
reachable at the same dotted path (verified by the new tests under
tests/ai/agent/mcp/).

Two places needed a lazy attribute lookup so existing monkeypatch tests
still reach the real call site after decomposition:

- finish_ai_block resolves tools_workflow._atomic_write_text via the
  package namespace at call time (test_finish_ai_block_io_error_*
  monkeypatches the package-level binding).
- The preview helpers resolve tools_inspection._MAX_PREVIEW_BYTES via
  the package namespace at call time (test_preview_data_tiff_cap
  monkeypatches the package-level binding).

scripts/check_god_files.py: removed the two files from
GOD_FILE_SIZE_WAIVERS (each new sub-module is below the 750 LOC
threshold); waiver list now shows 8 tracked + 0 NEW violations.

Tests: 24 new surface-preservation + per-sub-module behavior tests under
tests/ai/agent/mcp/; full existing tests/ai/ + tests/integration MCP set
remains green (153 passed, 92 pre-existing S40a skel skips).

Gate-Record: .workflow/records/1431-mcp-tools-pair.json
Task-Kind: refactor
Issue: #1431
