+++ .workflow/records/1431-mcp-tools-pair.json
@@ -1,0 +1,178 @@
+{
+  "admin_labels": [],
+  "amendments": [],
+  "branch": "refactor/issue-1431/mcp-tools-pair",
+  "changed_test_paths": [
+    "tests/ai/agent/mcp/test_tools_workflow_surface.py",
+    "tests/ai/agent/mcp/test_tools_inspection_surface.py"
+  ],
+  "check_results": [
+    {
+      "command_or_tool": "ruff check src/scistudio/ai/agent/mcp/tools_workflow/ src/scistudio/ai/agent/mcp/tools_inspection/ tests/ai/agent/mcp/ scripts/check_god_files.py",
+      "exit_code": 0,
+      "name": "ruff",
+      "output_path": "All checks passed!",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "ruff format --check src/scistudio/ai/agent/mcp/tools_workflow/ src/scistudio/ai/agent/mcp/tools_inspection/ tests/ai/agent/mcp/ scripts/check_god_files.py",
+      "exit_code": 0,
+      "name": "format",
+      "output_path": "17 files already formatted",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "PYTHONPATH=src python -m pytest tests/ai/ tests/integration/test_phase2_mcp_end_to_end.py tests/cli/test_mcp_bridge.py --timeout=60 --no-cov",
+      "exit_code": 0,
+      "name": "pytest",
+      "output_path": "153 passed, 92 skipped (pre-existing S40a skel markers)",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "python scripts/check_god_files.py --enforce",
+      "exit_code": 0,
+      "name": "god_file_advisory",
+      "output_path": "threshold=750 LOC; 8 tracked waivers (was 10), 0 NEW violations; tools_workflow.py + tools_inspection.py removed",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "python -m scistudio.qa.audit.full_audit --repo-root . --format json --output docs/audit/full-audit-latest.json",
+      "exit_code": 0,
+      "name": "full_audit",
+      "output_path": "docs/audit/full-audit-latest.json",
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    }
+  ],
+  "commit": null,
+  "docs_landing": {
+    "adr": {
+      "not_applicable": true,
+      "rationale": "N/A pure structural refactor; no architectural change"
+    },
+    "changelog": {
+      "not_applicable": true,
+      "rationale": "N/A umbrella #1427 lands rolled-up CHANGELOG entry after all Phase 1 sub-PRs merge"
+    },
+    "checklist": {},
+    "docs": {
+      "not_applicable": true,
+      "rationale": "N/A new module map documented in tools_workflow/__init__.py + tools_inspection/__init__.py docstrings; no external user-facing doc affected"
+    },
+    "spec": {
+      "not_applicable": true,
+      "rationale": "N/A no contract or schema change"
+    }
+  },
+  "full_audit": {
+    "blocks_merge": false,
+    "command": "python -m scistudio.qa.audit.full_audit --repo-root . --format json --output docs/audit/full-audit-latest.json",
+    "exit_code": 0,
+    "known_debt": [],
+    "output_path": "docs/audit/full-audit-latest.json",
+    "status": "pass",
+    "summary": {},
+    "unclassified_failures": []
+  },
+  "governance_touch": false,
+  "issues": [
+    {
+      "close_in_pr": true,
+      "followup_rationale": null,
+      "number": 1431,
+      "url": null
+    }
+  ],
+  "owner_directive": "Refactor MCP tools_workflow + tools_inspection into sub-packages per umbrella #1427",
+  "planned_files": [
+    "src/scistudio/ai/agent/mcp/tools_workflow/**",
+    "src/scistudio/ai/agent/mcp/tools_workflow.py",
+    "src/scistudio/ai/agent/mcp/tools_inspection/**",
+    "src/scistudio/ai/agent/mcp/tools_inspection.py",
+    "scripts/check_god_files.py"
+  ],
+  "pull_request": null,
+  "record_path": ".workflow/records/1431-mcp-tools-pair.json",
+  "required_checks": [
+    "ruff",
+    "format",
+    "pytest",
+    "god_file_advisory",
+    "full_audit",
+    "sentrux"
+  ],
+  "schema_version": "1",
+  "scope": {
+    "exclude": [],
+    "include": [
+      "src/scistudio/ai/agent/mcp/tools_workflow/**",
+      "src/scistudio/ai/agent/mcp/tools_workflow.py",
+      "src/scistudio/ai/agent/mcp/tools_inspection/**",
+      "src/scistudio/ai/agent/mcp/tools_inspection.py",
+      "tests/ai/agent/mcp/**",
+      "scripts/check_god_files.py"
+    ]
+  },
+  "sentrux": {
+    "command_or_tool": "sentrux check",
+    "mode": "free-tier",
+    "name": "sentrux.free_tier",
+    "output_path": "scan: files=1166 edges=2641 lines=299555 quality_signal=4444; check_rules: pass=true 0 violations rules_checked=3/15; health: bottleneck=acyclicity modularity=0.176 redundancy=0.085; session_end: signal_delta=0 stable, cycles_change=[3,3]",
+    "pro_required": false,
+    "quality_signal": null,
+    "rules_checked": null,
+    "status": "pass",
+    "summary": {},
+    "thresholds": {},
+    "total_rules_defined": null
+  },
+  "stages": [
+    {
+      "completed_at": null,
+      "stage": "scope_and_issue",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "plan",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "implement",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "update_docs",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "test_and_checks",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "commit_and_submit_pr",
+      "status": "done",
+      "summary": null
+    }
+  ],
+  "task_id": "1431-mcp-tools-pair",
+  "task_kind": "refactor"
+}
