+++ src/scistudio/ai/agent/mcp/tools_workflow/write.py
@@ -1,0 +1,187 @@
+"""Write-class workflow tools (3 of 4 — ``finish_ai_block`` lives in its own module).
+
+Tools: ``write_workflow``, ``run_workflow``, ``cancel_run``.
+
+Extracted from the original single-file ``tools_workflow.py`` (#1431,
+umbrella #1427). No behavior change.
+"""
+
+from __future__ import annotations
+
+import json
+import logging
+
+import yaml as yaml_module
+from filelock import FileLock, Timeout
+from pydantic import Field, ValidationError
+
+from scistudio.ai.agent.mcp._context import _resolve_project_path
+from scistudio.ai.agent.mcp.server import mcp
+from scistudio.ai.agent.mcp.tools_workflow._errors import (
+    _ensure_error_subscriber,
+    _run_block_errors,
+)
+from scistudio.ai.agent.mcp.tools_workflow._helpers import (
+    _LOCK_TIMEOUT_SECONDS,
+    _atomic_write_text,
+    _diff_summary,
+    _get_workflow_runtime,
+)
+from scistudio.ai.agent.mcp.tools_workflow._models import (
+    CancelRunResult,
+    RunWorkflowResult,
+    WriteWorkflowResult,
+)
+
+logger = logging.getLogger(__name__)
+
+
+# ---------------------------------------------------------------------------
+# (a.6) write_workflow  (write-class)
+# ---------------------------------------------------------------------------
+
+
+@mcp.tool(name="write_workflow", tags={"category:workflow", "write"})
+async def write_workflow(
+    path: str = Field(description="Project-relative path under workflows/."),
+    yaml: str = Field(description="Full workflow YAML content; will be schema-validated before write."),
+) -> WriteWorkflowResult:
+    """Persist a workflow YAML to disk with a file lock and pre-write schema validation.
+
+    Use when:
+      - You're creating a new workflow or replacing an existing one.
+
+    Do NOT use to:
+      - Patch one block's config (use ``update_block_config`` — preserves
+        comments and key order via ruamel.yaml round-trip).
+      - Edit ``workflows/*.yaml`` via Bash/Edit/Write tools — the
+        protect_workflow_yaml hook (ADR-040 §3.6) will block such calls.
+        This tool is the ONLY supported write path for workflows.
+
+    Returns ``WriteWorkflowResult`` with ``next_step`` pointing at
+    ``validate_workflow`` for canonical post-write verification.
+    """
+    from scistudio.workflow.schema import WorkflowFileModel
+
+    # Pre-write validation: parse YAML and run through the same pydantic
+    # model the runtime + GET route use. Failure raises ValueError with
+    # JSON-embedded structured pydantic errors.
+    try:
+        parsed = yaml_module.safe_load(yaml)
+    except yaml_module.YAMLError as exc:
+        raise ValueError(f"write_workflow: YAML parse failure: {exc}") from exc
+    try:
+        WorkflowFileModel.model_validate(parsed)
+    except ValidationError as exc:
+        raise ValueError(
+            "write_workflow: refusing to write — workflow does not match "
+            "the SciStudio schema. Errors (JSON):\n" + json.dumps(exc.errors(), indent=2, default=str)
+        ) from exc
+
+    p = _resolve_project_path(path)
+    lock_path = str(p) + ".lock"
+    try:
+        with FileLock(lock_path, timeout=_LOCK_TIMEOUT_SECONDS):
+            old = p.read_text(encoding="utf-8") if p.exists() else ""
+            bytes_written = _atomic_write_text(p, yaml)
+            summary = _diff_summary(old, yaml)
+    except Timeout as exc:
+        raise TimeoutError(
+            f"write_workflow: could not acquire lock for {p} within {_LOCK_TIMEOUT_SECONDS}s (someone else is editing?)"
+        ) from exc
+
+    logger.info("write_workflow: wrote %s (%s)", p, summary)
+    return WriteWorkflowResult(path=str(p), bytes_written=bytes_written, diff_summary=summary)
+
+
+# ---------------------------------------------------------------------------
+# (a.7) run_workflow  (write-class)
+# ---------------------------------------------------------------------------
+
+
+@mcp.tool(name="run_workflow", tags={"category:workflow", "write"})
+async def run_workflow(
+    path: str = Field(description="Project-relative path to the workflow YAML to execute."),
+) -> RunWorkflowResult:
+    """Submit a workflow for execution and return its run identifier.
+
+    Use when:
+      - You've written/validated a workflow and want to execute it.
+
+    Do NOT use to:
+      - Re-run a previously failed run with code fixes — write the fix
+        first (``write_workflow`` or block source edit + ``reload_blocks``),
+        then call this.
+      - Inspect run progress — poll ``get_run_status`` with the returned
+        ``run_id``.
+
+    Returns immediately with status='queued'. Progress is observable via
+    ``get_run_status``.
+    """
+    runtime = _get_workflow_runtime()
+    _ensure_error_subscriber()
+    resolved = _resolve_project_path(path)
+    workflow_id = resolved.stem
+    # Clear stale errors from a prior failed run with the same id.
+    for key in list(_run_block_errors.keys()):
+        if key[0] == workflow_id:
+            del _run_block_errors[key]
+    result = runtime.start_workflow(workflow_id)
+    run_id = result.get("workflow_id", workflow_id) if isinstance(result, dict) else workflow_id
+    logger.info("run_workflow: started run %s for %s", run_id, resolved)
+    return RunWorkflowResult(run_id=str(run_id), status="queued")
+
+
+# ---------------------------------------------------------------------------
+# (a.8) cancel_run  (write-class)
+# ---------------------------------------------------------------------------
+
+
+@mcp.tool(name="cancel_run", tags={"category:workflow", "write"})
+async def cancel_run(
+    run_id: str = Field(description="Identifier returned by run_workflow."),
+) -> CancelRunResult:
+    """Request cancellation of an in-flight workflow run.
+
+    Use when:
+      - A run is producing wrong results and you want to stop it early.
+      - A run is hung and needs to be terminated.
+
+    Do NOT use to:
+      - Inspect run state — use ``get_run_status``.
+
+    Raises ``KeyError`` if the run_id is unknown.
+    """
+    import asyncio
+
+    from scistudio.engine.events import CANCEL_WORKFLOW_REQUEST, EngineEvent
+
+    runtime = _get_workflow_runtime()
+    runs = getattr(runtime, "workflow_runs", None)
+    if not isinstance(runs, dict) or run_id not in runs:
+        raise KeyError(f"Unknown run: {run_id}")
+
+    run = runs[run_id]
+    event_bus = getattr(run.scheduler, "_event_bus", None) if hasattr(run, "scheduler") else None
+    if event_bus is None:
+        if hasattr(run, "task") and not run.task.done():
+            run.task.cancel()
+        cancel_requested = True
+    else:
+        coro = event_bus.emit(EngineEvent(event_type=CANCEL_WORKFLOW_REQUEST, data={"workflow_id": run_id}))
+        try:
+            loop = asyncio.get_running_loop()
+            run._cancel_task = loop.create_task(coro)  # type: ignore[attr-defined]
+        except RuntimeError:
+            await coro
+        cancel_requested = True
+
+    logger.info("cancel_run: requested cancellation for %s", run_id)
+    return CancelRunResult(run_id=run_id, cancel_requested=cancel_requested)
+
+
+__all__: list[str] = [
+    "cancel_run",
+    "run_workflow",
+    "write_workflow",
+]
